+++ lib/reactive_state.dart
@@ -3,4 +3,7 @@ library reactive_state;
 export 'src/autobuild.dart';
 export 'src/autorun.dart';
 export 'src/derived_value.dart';
+export 'src/list_value.dart';
+export 'src/map_value.dart';
+export 'src/observable_container.dart';
 export 'src/value.dart';